Fix mistake in API of `rescuing`: use effects
It should take a monadic function, i.e. a `Block(e -> Eff<a>)`,
instead of a pretend-pure function, i.e. `Block(e -> a)`.
otherwise it wouldn't be composable with other effects.

If returning a pure value, as demonstrated in the tests added in this
commit, just wrap your pure value (e.g. `42`) with `Eff.pure`.